refactor(frontend): move QrCode step to SendModal
