+++ src/frontend/src/lib/components/send/SendModal.svelte
@@ -1,7 +1,11 @@
 <script lang="ts">
 	import { WizardModal, type WizardStep, type WizardSteps } from '@dfinity/gix-components';
 	import { createEventDispatcher, setContext } from 'svelte';
+	import { enabledErc20Tokens } from '$eth/derived/erc20.derived';
+	import { enabledEthereumTokens } from '$eth/derived/tokens.derived';
+	import { decodeQrCode as decodeQrCodeETH } from '$eth/utils/qr-code.utils';
 	import { icrcAccountIdentifierText } from '$icp/derived/ic.derived';
+	import SendQRCodeScan from '$lib/components/send/SendQRCodeScan.svelte';
 	import SendTokenContext from '$lib/components/send/SendTokenContext.svelte';
 	import SendTokensList from '$lib/components/send/SendTokensList.svelte';
 	import SendWizard from '$lib/components/send/SendWizard.svelte';
@@ -23,8 +27,11 @@
 	} from '$lib/stores/modal-tokens-list.store';
 	import { token } from '$lib/stores/token.store';
 	import type { Network, NetworkId } from '$lib/types/network';
-	import type { Token } from '$lib/types/token';
+	import type { QrResponse, QrStatus } from '$lib/types/qr-code';
+	import type { OptionToken, Token } from '$lib/types/token';
 	import { closeModal } from '$lib/utils/modal.utils';
+	import { isNetworkIdEthereum } from '$lib/utils/network.utils';
+	import { decodeQrCode } from '$lib/utils/qr-code.utils';
 	import { goToWizardStep } from '$lib/utils/wizard-modal.utils';
 
 	export let destination = '';
@@ -94,6 +101,29 @@
 			stepName
 		});
 
+	const onDecodeQrCode = ({
+		status,
+		code,
+		expectedToken
+	}: {
+		status: QrStatus;
+		code?: string;
+		expectedToken: OptionToken;
+	}): QrResponse => {
+		const params = {
+			status,
+			code,
+			expectedToken
+		};
+		return isNetworkIdEthereum($token?.network.id)
+			? decodeQrCodeETH({
+					...params,
+					ethereumTokens: $enabledEthereumTokens,
+					erc20Tokens: $enabledErc20Tokens
+				})
+			: decodeQrCode(params);
+	};
+
 	// TODO: Use network id to get the address to support bitcoin.
 	let source: string;
 	$: source = $icrcAccountIdentifierText ?? '';
@@ -118,6 +148,19 @@
 			/>
 		{:else if currentStep?.name === WizardStepsSend.FILTER_NETWORKS}
 			<ModalNetworksFilter on:icNetworkFilter={() => goToStep(WizardStepsSend.TOKENS_LIST)} />
+		{:else if currentStep?.name === WizardStepsSend.QR_CODE_SCAN}
+			<SendQRCodeScan
+				expectedToken={$token}
+				bind:destination
+				bind:amount
+				decodeQrCode={onDecodeQrCode}
+				on:icQRCodeBack={() =>
+					goToWizardStep({
+						modal,
+						steps,
+						stepName: WizardStepsSend.SEND
+					})}
+			/>
 		{:else}
 			<SendWizard
 				{source}
@@ -137,12 +180,6 @@
 						modal,
 						steps,
 						stepName: WizardStepsSend.QR_CODE_SCAN
-					})}
-				on:icQRCodeBack={() =>
-					goToWizardStep({
-						modal,
-						steps,
-						stepName: WizardStepsSend.SEND
 					})}
 			/>
 		{/if}
