+++ src/frontend/src/lib/components/send/SendWizard.svelte
@@ -47,7 +47,6 @@
 		on:icNext
 		on:icClose
 		on:icQRCodeScan
-		on:icQRCodeBack
 	/>
 {:else if isNetworkIdEvm($sendToken.network.id) && nonNullish($selectedEvmNetwork) && nonNullish($evmNativeToken)}
 	<!--			TODO: use store evmNativeToken here when we adapt the fee context to fetch the EVM fees -->
@@ -65,7 +64,6 @@
 		on:icNext
 		on:icClose
 		on:icQRCodeScan
-		on:icQRCodeBack
 	/>
 {:else if isNetworkIdICP($sendToken.network.id)}
 	<IcSendTokenWizard
@@ -81,7 +79,6 @@
 		on:icNext
 		on:icClose
 		on:icQRCodeScan
-		on:icQRCodeBack
 	/>
 {:else if isNetworkIdBitcoin($sendToken.network.id)}
 	<BtcSendTokenWizard
@@ -95,7 +92,6 @@
 		on:icClose
 		on:icSendBack
 		on:icQRCodeScan
-		on:icQRCodeBack
 	/>
 {:else if isNetworkIdSolana($sendToken.network.id)}
 	<SolSendTokenWizard
@@ -109,7 +105,6 @@
 		on:icClose
 		on:icSendBack
 		on:icQRCodeScan
-		on:icQRCodeBack
 	/>
 {:else}
 	<slot />
